+++ lib/src/enums/chord_type.dart
